Clean up gpg/cli
This commit rearranges the GPG CLI implementation a bit to make things
easier to find and maintain. Effectively it should be a no-op.

RELEASE_NOTES=n/a
