+++ internal/backend/crypto/gpg/cli/gpg.go
@@ -2,16 +2,12 @@
 package cli
 
 import (
-	"bufio"
-	"bytes"
 	"context"
 	"os"
-	"os/exec"
-	"strings"
 
 	"github.com/gopasspw/gopass/internal/backend/crypto/gpg"
-	"github.com/gopasspw/gopass/internal/out"
 	"github.com/gopasspw/gopass/pkg/debug"
+	"github.com/gopasspw/gopass/pkg/fsutil"
 	lru "github.com/hashicorp/golang-lru"
 )
 
@@ -58,7 +54,7 @@ func New(ctx context.Context, cfg Config) (*GPG, error) {
 	g := &GPG{
 		binary:    "gpg",
 		args:      append(defaultArgs, cfg.Args...),
-		throwKids: fileContains(gpgConfigLoc(), "throw-keyids"),
+		throwKids: fsutil.FileContains(gpgConfigLoc(), "throw-keyids"),
 	}
 
 	debug.Log("initializing LRU cache")
@@ -78,89 +74,6 @@ func New(ctx context.Context, cfg Config) (*GPG, error) {
 	debug.Log("binary detected")
 
 	return g, nil
-}
-
-// RecipientIDs returns a list of recipient IDs for a given file
-func (g *GPG) RecipientIDs(ctx context.Context, buf []byte) ([]string, error) {
-	_ = os.Setenv("LANGUAGE", "C")
-	recp := make([]string, 0, 5)
-
-	args := []string{"--batch", "--list-only", "--list-packets", "--no-default-keyring", "--secret-keyring", "/dev/null"}
-	cmd := exec.CommandContext(ctx, g.binary, args...)
-	cmd.Stdin = bytes.NewReader(buf)
-	debug.Log("%s %+v", cmd.Path, cmd.Args)
-
-	cmdout, err := cmd.CombinedOutput()
-	if err != nil {
-		return []string{}, err
-	}
-
-	scanner := bufio.NewScanner(bytes.NewBuffer(cmdout))
-	for scanner.Scan() {
-		line := strings.TrimSpace(scanner.Text())
-		debug.Log("GPG Output: %s", line)
-		if !strings.HasPrefix(line, ":pubkey enc packet:") {
-			continue
-		}
-		m := splitPacket(line)
-		if keyid, found := m["keyid"]; found {
-			kl, err := g.listKeys(ctx, "public", keyid)
-			if err != nil || len(kl) < 1 {
-				continue
-			}
-			recp = append(recp, kl[0].Fingerprint)
-		}
-	}
-
-	if g.throwKids {
-		// TODO shouldn't log here
-		out.Warningf(ctx, "gpg option throw-keyids is set. some features might not work.")
-	}
-	return recp, nil
-}
-
-// Encrypt will encrypt the given content for the recipients. If alwaysTrust is true
-// the trust-model will be set to always as to avoid (annoying) "unusable public key"
-// errors when encrypting.
-func (g *GPG) Encrypt(ctx context.Context, plaintext []byte, recipients []string) ([]byte, error) {
-	args := append(g.args, "--encrypt")
-	if gpg.IsAlwaysTrust(ctx) {
-		// changing the trustmodel is possibly dangerous. A user should always
-		// explicitly opt-in to do this
-		args = append(args, "--trust-model=always")
-	}
-	for _, r := range recipients {
-		kl, err := g.listKeys(ctx, "public", r)
-		if err != nil {
-			debug.Log("Failed to check key %s. Adding anyway. %s", err)
-		} else if len(kl.UseableKeys(gpg.IsAlwaysTrust(ctx))) < 1 {
-			out.Printf(ctx, "Not using expired key %s for encryption", r)
-			continue
-		}
-		args = append(args, "--recipient", r)
-	}
-
-	buf := &bytes.Buffer{}
-
-	cmd := exec.CommandContext(ctx, g.binary, args...)
-	cmd.Stdin = bytes.NewReader(plaintext)
-	cmd.Stdout = buf
-	cmd.Stderr = os.Stderr
-
-	debug.Log("%s %+v", cmd.Path, cmd.Args)
-	err := cmd.Run()
-	return buf.Bytes(), err
-}
-
-// Decrypt will try to decrypt the given file
-func (g *GPG) Decrypt(ctx context.Context, ciphertext []byte) ([]byte, error) {
-	args := append(g.args, "--decrypt")
-	cmd := exec.CommandContext(ctx, g.binary, args...)
-	cmd.Stdin = bytes.NewReader(ciphertext)
-	cmd.Stderr = os.Stderr
-
-	debug.Log("%s %+v", cmd.Path, cmd.Args)
-	return cmd.Output()
 }
 
 // Initialized always returns nil
