--- internal/backend/crypto/gpg/cli/import_others_test.go
@@ -1,22 +1,0 @@
-// +build !windows
-
-package cli
-
-import (
-	"context"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestImport(t *testing.T) {
-	ctx := context.Background()
-
-	g := &GPG{}
-	g.binary = "true"
-
-	assert.NoError(t, g.ImportPublicKey(ctx, []byte("foobar")))
-
-	g.binary = ""
-	assert.Error(t, g.ImportPublicKey(ctx, []byte("foobar")))
-}
