--- internal/backend/crypto/gpg/cli/umask_others.go
@@ -1,9 +1,0 @@
-// +build !windows
-
-package cli
-
-import "syscall"
-
-func umask(mask int) int {
-	return syscall.Umask(mask)
-}
