--- internal/backend/crypto/gpg/cli/umask_windows.go
@@ -1,7 +1,0 @@
-// +build windows
-
-package cli
-
-func umask(mask int) int {
-	return -1
-}
