+++ internal/backend/crypto/gpg/cli/utils.go
@@ -2,34 +2,9 @@ package cli
 
 import (
 	"os"
-	"strconv"
+	"path/filepath"
 	"strings"
-	"time"
 )
-
-// parseTS parses the passed string as an Epoch int and returns
-// the time struct or the zero time struct
-func parseTS(str string) time.Time {
-	t := time.Time{}
-
-	if sec, err := strconv.ParseInt(str, 10, 64); err == nil {
-		t = time.Unix(sec, 0)
-	}
-
-	return t
-}
-
-// parseInt parses the passed string as an int and returns it
-// or 0 on errors
-func parseInt(str string) int {
-	i := 0
-
-	if iv, err := strconv.ParseInt(str, 10, 32); err == nil {
-		i = int(iv)
-	}
-
-	return i
-}
 
 func splitPacket(in string) map[string]string {
 	m := make(map[string]string, 3)
@@ -42,6 +17,16 @@ func splitPacket(in string) map[string]string {
 		m[p[i]] = strings.Trim(p[i+1], ",")
 	}
 	return m
+}
+
+// gpgConfigLoc returns the location of the GPG config file
+func gpgConfigLoc() string {
+	if sv := os.Getenv("GNUPGHOME"); sv != "" {
+		return filepath.Join(sv, "gpg.conf")
+	}
+
+	uhd, _ := os.UserHomeDir()
+	return filepath.Join(uhd, ".gnupg", "gpg.conf")
 }
 
 // GPGOpts parses extra GPG options from the environment
