+++ internal/backend/crypto/gpg/cli/utils_windows.go
@@ -5,3 +5,7 @@ package cli
 func tty() string {
 	return ""
 }
+
+func umask(mask int) int {
+	return -1
+}
