+++ internal/backend/crypto/gpg/colons/parse_colons_test.go
@@ -1,4 +1,4 @@
-package cli
+package colons
 
 import (
 	"strings"
