+++ internal/backend/crypto/gpg/colons/parse_fuzz.go
@@ -1,11 +1,11 @@
 // +build gofuzz
 
-package cli
+package colons
 
 import "bytes"
 
 func Fuzz(data []byte) int {
-	if kl := parseColons(bytes.NewReader(data)); len(kl) != 0 {
+	if kl := Parse(bytes.NewReader(data)); len(kl) != 0 {
 		return 1
 	}
 	return 0
