+++ internal/backend/crypto/gpg/colons/utils.go
@@ -1,0 +1,30 @@
+package colons
+
+import (
+	"strconv"
+	"time"
+)
+
+// parseTS parses the passed string as an Epoch int and returns
+// the time struct or the zero time struct
+func parseTS(str string) time.Time {
+	t := time.Time{}
+
+	if sec, err := strconv.ParseInt(str, 10, 64); err == nil {
+		t = time.Unix(sec, 0)
+	}
+
+	return t
+}
+
+// parseInt parses the passed string as an int and returns it
+// or 0 on errors
+func parseInt(str string) int {
+	i := 0
+
+	if iv, err := strconv.ParseInt(str, 10, 32); err == nil {
+		i = int(iv)
+	}
+
+	return i
+}
